auth and 4 functions

+++ server.js
@@ -1,22 +1,21 @@
-// if req.url = '/' make it point to index
-
 const http = require('http');
 const fs = require('fs');
 const path = require('path');
 const querystring = require('querystring');
-var elemCount = 2;
+const template = '';
+var numOfElements = 2;
 
 const server = http.createServer((req, res) =>  {
   if (req.url === '/') {
-    req.url = '/index.html'; // slash b4 index?
+    req.url = '/index.html';
   }
   if(req.method === 'GET'){
     getFunction(req, res);
-  } else if(req.method === 'POST'){
+  } else if(req.method === 'POST' && pswdCheck(req, res)){
     postFunction(req, res);
-  } else if (req.method === 'DELETE') {
+  } else if (req.method === 'DELETE' && pswdCheck(req, res)) {
     deletePost(req, res);
-  } else if (req.method === 'PUT') {
+  } else if (req.method === 'PUT' && pswdCheck(req, res)) {
     postFunction(req, res);
   }
 });
@@ -39,35 +38,45 @@ const postFunction = (req, res) => {
     const reqBody = querystring.parse(data.toString());
     addNewElem(req, reqBody);
     fs.writeFileSync('public' + req.url, htmlTemplate(reqBody));
-/*    fs.readFile('./public' + req.url, (err, data) => {
-      console.log(err + "  " + data);
-    });*/
     res.end();
   });
 };
 function addNewElem(req, reqBody) {
   fs.readFile('./public' + req.url, (err, data) => {
     if(err !== null) {
-      console.log('writing started');
       var x = fs.readFileSync('public/index.html').toString();
       var y = x.slice(0, x.search('</ol>') - 3);
       var z = x.slice(x.search('</ol>') - 3, x.length);
       x = y + `\r\n    <li>\r\n      <a href="${req.url}">${reqBody.elementName}</a>\r\n    </li>` + z;
+
+      let findTheNum = x.indexOf(`</h3>`);
+      numOfElements = parseFloat(x.charAt(findTheNum-1));
+      let incrementNumElements = ++numOfElements;
+      let htmlArray = x.split('\n');
+      htmlArray.splice(10,1,`<h3>There are ${incrementNumElements}</h3>`);
+      x = htmlArray.join(`\n`);
       fs.writeFileSync('public/index.html', x);
-      // find the 2 and replace it... google how to do this
-      x.charAt(x.search(' ' + elemCount) + 1) = elemCount + 1;
-      elemCount++;
     }
   });
 }
 function deletePost(req, res) {
   var x = fs.readFileSync('public/index.html').toString();
-  var z = x.slice(x.search('<a href=\"' + req.url) + req.url.length + 15 + (req.url.length + 6), x.length);
+  var z = x.slice(x.search('<a href=\"' + req.url) + (req.url.length * 2) + 21, x.length);
   var y = x.slice(0, x.search('<a href=\"' + req.url) - 12);
   x = y + z;
   fs.writeFileSync('public/index.html', x);
   fs.unlinkSync('./public' + req.url);
+  numOfElements--;
   res.end();
+}
+function pswdCheck(req, res) {
+  if (req.headers.pswd === 'abc123') {
+    console.log('accepted');
+    return true;
+  } else {
+    console.log('denied');
+    res.end();
+  }
 }
 
 const htmlTemplate = (reqBody) => (
